ml-service/requirements.txt
# ML Service Dependencies
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.4
pandas==2.2.0
scikit-learn==1.4.0
python-dotenv==1.0.0
xgboost==2.0.3
httpx==0.26.0
